Split RepositoryConfig
